// src/models/Employee.ts
export interface Employee {
  emp_id?: number;
  full_name: string;
  name_initial: string;
  display_name: string;
  gender_id?: number;
  date_of_birth: string;
  email: string;
  mobile_no?: number;
  designation: string;
  emp_type?: number;
  join_date: string;
  experience?: number;
  salary?: number;
  personal_notes: string;
}